script done

+++ csci-340/Homework/hw1/hw1-commands.sh
@@ -1,13 +1,25 @@
 #!bin/bash
 
+# validate script is run properly
+if [ $# -lt 1 ]; then
+    echo "Must provide an output file as argument."
+    exit
+fi
+if [ $1 == $0 ]; then
+    echo "Output file must be different from script file."
+    exit
+fi
+printf "Script executing..."
+
+# designate output file
 OUT_FILE=$1
 
 # (re)create output file
 rm $OUT_FILE
 touch $OUT_FILE
 
+# header
 printf "Collin Bauer\nCSCI 340 - Operating Systems\nDr. Crosby\nSpring 2020\n\n" >> $OUT_FILE
-
 echo "Note that, while I do have a Ubuntu VM on my Windows laptop, I also have a Linux laptop." >> $OUT_FILE
 echo "I plan to do most coding from my Linux laptop, but test sensitive code on the VM." >> $OUT_FILE
 printf "\nThe following output is from my Linux laptop.\n" >> $OUT_FILE
@@ -31,3 +43,6 @@ echo "ps (short for processes) outputs a list of currently running processes." >
 echo "  -l makes the output take a long-hand form, reporting more information like parent process ID and memory address." >> $OUT_FILE
 echo "  -a extends the list to include all running processes, not just those visible to the bash shell." >> $OUT_FILE
 echo "  -al combines the functionality of both above arguments." >> $OUT_FILE
+
+printf "Done.\n"
+exit
